requirements.txt
pandas
numpy
PyMuPDF
openai
deepl
python-dotenv
quart
pdfplumber
